requirements.txt
aiohttp==3.9.5
aiosignal==1.3.1
annotated-types==0.7.0
anyio==4.4.0
attrs==23.2.0
azure-common==1.1.28
azure-core==1.30.2
azure-identity==1.17.1
azure-mgmt-authorization==4.0.0
azure-mgmt-core==1.4.0
azure-mgmt-subscription==3.1.1
certifi==2024.7.4
cffi==1.16.0
charset-normalizer==3.3.2
cryptography==42.0.8
Deprecated==1.2.14
frozenlist==1.4.1
h11==0.14.0
h2==4.1.0
hpack==4.0.0
httpcore==1.0.5
httpx==0.27.0
hyperframe==6.0.1
idna==3.7
importlib_metadata==7.1.0
isodate==0.6.1
isort==5.13.2
microsoft-kiota-abstractions==1.3.3
microsoft-kiota-authentication-azure==1.0.0
microsoft-kiota-http==1.3.2
microsoft-kiota-serialization-form==0.1.0
microsoft-kiota-serialization-json==1.2.0
microsoft-kiota-serialization-multipart==0.1.0
microsoft-kiota-serialization-text==1.0.0
msal==1.29.0
msal-extensions==1.2.0
msgraph-core==1.1.1
msgraph-sdk==1.5.2
msrest==0.7.1
multidict==6.0.5
neo4j==5.22.0
oauthlib==3.2.2
opentelemetry-api==1.25.0
opentelemetry-sdk==1.25.0
opentelemetry-semantic-conventions==0.46b0
pendulum==3.0.0
portalocker==2.10.0
pycparser==2.22
pydantic==2.8.2
pydantic_core==2.20.1
PyJWT==2.8.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.1
requests==2.32.3
requests-oauthlib==2.0.0
setuptools==69.5.1
six==1.16.0
sniffio==1.3.1
std-uritemplate==1.0.3
time-machine==2.14.2
typing_extensions==4.12.2
tzdata==2024.1
urllib3==2.2.2
wheel==0.43.0
wrapt==1.16.0
yarl==1.9.4
zipp==3.19.2
